requirement.txt
asgiref==3.5.2
autopep8==1.7.0       
dj-static==0.0.6      
Django==4.0.6
django-autoslug==1.9.8
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-js-asset==2.0.0
django-quill-editor==0.1.40
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tzdata==2022.1